app/requirements.txt
google-cloud-storage==1.44.0
google-cloud-firestore==2.6.0